+++ MinkowskiEngine/utils/__init__.py
@@ -21,7 +21,7 @@
 # Please cite "4D Spatio-Temporal ConvNets: Minkowski Convolutional Neural
 # Networks", CVPR'19 (https://arxiv.org/abs/1904.08755) if you use any part
 # of the code.
-from .voxelization import sparse_quantize, ravel_hash_vec, fnv_hash_vec
+from .quantization import sparse_quantize, ravel_hash_vec, fnv_hash_vec
 from .collation import SparseCollation, batched_coordinates, sparse_collate
 from .gradcheck import gradcheck
 from .init import kaiming_normal_