+++ tests/quantization.py
@@ -25,6 +25,7 @@ import unittest
 import numpy as np
 
 from MinkowskiEngine.utils import sparse_quantize
+import MinkowskiEngineBackend as MEB
 
 
 class TestGPUVoxelization(unittest.TestCase):
@@ -46,6 +47,46 @@ class TestGPUVoxelization(unittest.TestCase):
         quantized_coords, quantized_feats, quantized_labels = sparse_quantize(
             coords, feats, labels, ignore_label)
         print(quantized_labels)
+
+    def test_mapping(self):
+        N = 16575
+        coords = (np.random.rand(N, 3) * 100).astype(np.int32)
+        mapping = MEB.quantize(coords)
+        print(mapping)
+        print('N unique:', len(mapping), 'N:', N)
+
+        print(sparse_quantize(coords))
+
+    def test_label(self):
+        N = 16575
+        ignore_label = 255
+
+        coords = (np.random.rand(N, 3) * 100).astype(np.int32)
+        feats = np.random.rand(N, 4)
+        labels = np.floor(np.random.rand(N) * 3)
+
+        labels = labels.astype(np.int32)
+
+        # Make duplicates
+        coords[:3] = 0
+        labels[:3] = 2
+
+        mapping, colabels = MEB.quantize_label(coords, labels, ignore_label)
+        print(mapping)
+        print(colabels)
+        print('Unique labels and counts:', np.unique(colabels, return_counts=True))
+        print('N unique:', len(mapping), 'N:', N)
+
+        print(sparse_quantize(coords, feats, labels, ignore_label))
+
+    def test_collision(self):
+        N = 4
+        coords = np.array([[0, 0], [0, 0], [0, 0], [0, 1]], dtype=np.int32)
+        labels = np.array([0,     1,      2,      3], dtype=np.int32)
+
+        mapping, colabels = sparse_quantize(coords, labels=labels, ignore_label=255)
+        print(mapping)
+        print(colabels)
 
 
 if __name__ == '__main__':